pyproject.toml
[project]
name = "dev"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.70.0",
    "google-genai>=1.45.0",
    "numpy>=2.3.4",
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "voyageai>=0.3.5",
    "xai-sdk>=1.3.0",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "ruff>=0.14.0",
]

[tool.mypy]
python_version = "3.12"

[[tool.mypy.overrides]]
module = "xai_sdk.*"
ignore_missing_imports = true
